[UPDATE] LEVEL-10 - Modified counterpart tabele and related actions

+++ database/migrations/2023_10_26_092129_create_medical_shares_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->foreignId('student_id')->constrained();
             $table->string('medical_concern');
-            $table->integer('total_cost');
-            $table->integer('amount_paid');
+            $table->decimal('total_cost');
+            $table->decimal('amount_paid');
             $table->date('date');
             $table->timestamps();
         });
